split off a bunch of funcitons into dataflow_extras
